chore: 🤖 add tools
