+++ .eslintrc.json
@@ -1,0 +1,25 @@
+{
+  "parser": "@typescript-eslint/parser",
+  "extends": [
+    "react-app",
+    "plugin:@typescript-eslint/recommended"
+  ],
+  "parserOptions": {
+    "ecmaVersion": 2018,
+    "sourceType": "module"
+  },
+  "rules": {
+    "no-console": 2,
+    "no-unused-vars": 2,
+    "max-len": [
+      "error",
+      {
+        "code": 80
+      }
+    ],
+    "quotes": [
+      "error",
+      "single"
+    ]
+  }
+}
