+++ package.json
@@ -4,17 +4,25 @@
   "private": true,
   "dependencies": {
     "@gooddata/react-components": "6.2.0",
+    "@types/classnames": "^2.2.10",
+    "@types/jest": "^26.0.14",
+    "@types/node": "^14.11.2",
+    "@types/react": "^16.9.50",
+    "@types/react-dom": "^16.9.8",
+    "@types/react-transition-group": "^4.4.0",
     "http-proxy-middleware": "0.19.1",
     "node-sass": "4.14.1",
     "react": "16.5.2",
     "react-dom": "16.5.2",
-    "react-scripts": "3.4.3"
+    "react-scripts": "3.4.3",
+    "typescript": "^4.0.3"
   },
   "scripts": {
     "start": "cross-env HTTPS=true react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "bump": "npx standard-version"
   },
   "browserslist": [
     ">0.2%",
@@ -23,6 +31,7 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "cross-env": "6.0.3"
+    "cross-env": "6.0.3",
+    "standard-version": "^9.0.0"
   }
 }
